LINE Bot v0.2.0/requirements.txt
Flask==2.0.1
gunicorn==20.1.0
line-bot-sdk==1.20.0
APScheduler==3.6.3
psycopg2==2.8.4
Flask-Login==0.5.0
flask-talisman==0.7.0
